Move ZenUtils.redis to ZenUtils.RedisUtils
ZenUtils/redis was hiding Python2.7/site-packages/redis when running
commands from the ZenUtils directory (e.g. ZenUtils/zenpack.py)
